third_party/blink/web_tests/platform/generic/fast/dom/gc-7-expected.txt

This test verifies that objects are protected against garbage collection as long as the document is alive (for contrast with test 6).

The output should be the following pieces of text on lines by themselves: "B", "[object HTMLSpanElement]", "[object HTMLSpanElement]", "[object HTMLDocument]".

B
[object HTMLSpanElement]
[object HTMLSpanElement]
[object HTMLDocument]
